Improve HP4C design with respect to setup: tset_control entries
The task is a bit more clear now.  Part of the previous confusion
stemmed from lack of clarity with respect to the roles of
tset_inspect_XX_YY tables vs. tset_control.  Conversation w/
Eric helped a lot in clarifying the task and resulted in another
evolution regarding pc.state.  Not only will the resulting
output be more optimal in terms of minimizing resubmits, but
production is easier to understand.