Progress with gen_tset_inspect_XX_YY: basics
Basics taken care of, like the parse_control state, the table name,
and population of key data structures required to generate the
rest.  Note that now TICS is a subclass of HP4_Command.  Next task
is to fill in the match parameters in walk_parse_tree under the
for selectopt... block.