Complete bug fixes for match params handling, tset_inspect_XX_YY
Big chunk of the task for tset_inspect_XX_YY entries looks to be
complete after having tested it on all demo P4 programs.  Should
add to the unit tests to prevent regressions in the future. This
was a bit of a beast in terms of details that required solving,
happy to be done with this part.