Fix gen_tset_pipeline_entries bug, simplify pytest sync issue
Created a build function in HP4C that calls all the functions,
eliminating the sync issue I was having with the pytests as
I add functions to the build process.

The gen_tset_pipeline_entries method was outputting
[EXTRACTED_EXACT] as the first action parameter even though
two of the test programs employ matching on fields of
different source types in the first table.  This has been
fixed and pytests corrected.