Fix bug with demo one, phase C: route packets between subnets!
The bug was a missing t_set_validbits entry, which is a table
that doesn't do anything useful for the routing function, but
since there was no default entry and the table is unconditionally
invoked, the result was dropped packet.

In the future we might do something different, such as only
apply the t_set_validbits table if a conditional evaluates
true, which could involve a state variable whose value is
set during the setup phase.  In the meantime, always set
the default action for t_set_validbits to _no_op.