Finalize branch for public use.
This commit cleans out directories and files not relevant to
those seeking to examine HyPer4 code and functionality.

It completes the README and places it in the top directory
for appropriate visibility.

It add a NOTICE with verbiage for the Apache 2.0 license,
which applies to all software in this repository.