Complete JSON output of entry templates for match tables
With this commit, the compiler produces two output files: the
usual output for loading the *function* into HyPer4, and now,
the entry templates for use with the DPMU when one wants to
populate the tables of the source program defining that function.

These entry templates are in JSON format and intended to be
loaded as HP4_Match_Command objects (we have yet to write
a conversion function for this purpose but it should be
trivial).