Clean up tstgXY_update_state; start design thoughts t_mod_XY
Clearly we need a data structure populated: a dictionary with
the action_ID as the key and a tuple as the value, in which
the first member is a list of stages from which the action
might be invoked, and the second is a list of tuples, each
representing a call to a primitive, with the type and
subtype.  This will help not only t_mod_XY, but also
for generating the entries for tstgXY_update_state.