Add design re: modify_field handling, update HP4 accordingly
While thinking through the way modify_field calls are handled,
I recognized a need to update HP4 such that for t_mod_XY
and t_math_on_field_XY, the match_ID match parameter should
use the ternary match type so we can turn it on or off as
needed.  The issue is that we are handling a modify_field
call with a constant as the source parameter the same
way as a modify_field call with an action parameter reference
as the source parameter.  In the first case, when it comes
time to issue the t_mod_XY entries, we would need a new entry
for every possible match_ID even though every other parameter
stays the same.  So in such cases we need to 'turn off' the
use of the match_ID in matching.  By making the match type
'ternary', we can just do 0&&&0.  And for the second case
where we SHOULD issue a separate entry for every match_ID,
we can do [val]&&&0x7FFFFF (mask is for the 23-bit match_ID
field).

So HP4 was updated and regenerated.  We have yet to generate
the entries for t_mod_XY or t_math_on_field but this commit
should pave the way for a sensible approach.