Populate data structure supporting action handling
The data structure is a dictionary with a p4_action as the
key and an Action_Rep object as the value.  The Action_Rep
object has members such as stages, a set of all stages
from which the action may be invoked, and call_sequence,
a list of tuples of the form (primitive type, primitive
subtype) representing each of the primitives in the action.