Fix and standardize HP4 primitive execution tables
Work on the compiler led to realization that HyPer4 was flawed
with respect to primitive emulation in several cases.  It is
now apparent that to handle any primitive that accepts parameters,
the corresponding HP4 table must include the action_ID and the
match_ID in the match parameters.  It is also more clear now that
in all cases we should use the ternary match type for the
match_ID so we can turn it on or off as needed.  It is finally
clear that to simplify the compiler's job, the sequence of
match parameters should be standardized across all such tables.

These things have been fixed and HP4 has been regenerated.