Progress, gen_action_aparams: mod_extracted_extracted done
Of the 14 currently defined modify_field primitive subtypes in
HP4, the compiler initially needs to support 8 of them in order
to meet the four demo programs' requirements.  This commit
completes support for the subtype in which the dest and source
parameters for the modify_field call are both fields from the
parsed representation ('extracted', living in extracted.data in
HP4).