Incremental progress on HP4 redesign.
Biggest outstanding issue right now: whether to do the ext-to-
extracted.data copy before or after the tset_parse_select_XX_YY
table(s).  If we do it before, it simplifies things greatly,
but may be more expensive to do, because we will do the copy
every time we resubmit the packet.  At the same time then we
need to try to parallelize this task.