Add design thoughts re: order of ext-to-extracted.data copy
This commit adds a basic evaluation plan to figure out whether
the ext-to-extracted.data copy should happen first thing in the
ingress pipeline, or after we know parsing is complete.

This commit also starts work on a python script to analyze
pcap for the latency of different systems we need to compare.