Start revamp of evalrun to use Scapy, simplifying pcap parsing
I had been doing the parsing manually and this was an ugly mess
because the text output of tcpdump -r changes based on the layers
present in the packet.  So even though tests looked fine with
ICMP, the code did not parse TCP packets correctly.

Discovered that Scapy can do pcap parsing rather well.