Fix gen_action_entries bug
Depending on how the P4 source was written, in some cases the
tstgXY_update_state entry for 'finish_action' was looking for
the next_stage action parameter but didn't have one because
there were no further tables after this invocation of the
action, triggering a KeyError on 'None'.  This commit
detects these cases and supplies '0' for the parameter.