DPMU design work: task list, physical port handling discussion
This commit provides a detailed task list and starts the discussion
about how physical ports are handled in various ways.  The reason
this is a concern is because physical ports are resources assigned
to users in the userfile, but the user/client is and should
remain oblivious to the actual physical ports assigned to his
instances, and the DPMU performs the mapping from the port number
viewed by the client to the port number assigned in the server.

There are several cases to worry about for the various matching
and primitives that may involve the use of physical ports.