Acknowledge need to address tset_context limitation
When packets are received by HyPer4, program assignment is done
via the tset_context table, which *currently* matches only on
the physical ingress port, but was always intended to be
modified so that it could match on a variety of metadata and
parsed representation fields and even stateful memory.

Just how to do this sensibly is an open question.  We consider
addressing it now because currently in the dpmu we are handling
context->instance enforcement; that is, if one 'load' command
results in the assignment of a particular program instance to
a particular context, future 'load' commands should not reuse
part or all of that context for another program instance
assignment.  Currently context simply includes the physical ingress
port and we've always said HP4 should support a much greater
variety of context definition, such as the flowspaces definition
in the FlowVisor paper.

Realistically however we should delay this for the time being
and proceed with context management dealing only with the
physical ingress port.