Fix tset_parse_select_XX_YY action params
Two issues fixed:
- The first param should be the number of bytes
to extract from the packet on the next resubmit, but we were
generating the number of bits.
- A priority is required as the last parameter because ternary
matching is involved.  For entries that are 'don't care' as far
as the parsed representation is concerned, we emit the max
priority possible; otherwise we do care and we emit priority 0
(taking precedence over all higher values).