Fix #3/3 p4c-hp4 bugs; priorities provided for mod entries
Uncovered a p4c-bmv2 bug though: bitwidths of action parameters
are prematurely set to that of the destination field of a
modify_field primitive in which they are involved in the source.
This causes incorrect/unexpected behavior, causing issues both
with adding rules to affected tables, and the results of
invoking such actions where we are able to add rules.

Could just edit the JSON, but should probably figure out a fix
for p4c-bmv2.