Edit bitwidth_bugs w/ enlightened understanding; no p4c-bm bug.
The modified file reflects a better understanding of how
modify_field is handled by p4c-bm when the width of the dest
field is narrower than that of the source.  I no longer
believe p4c-bm has a bug; rather, I need to change p4c-hp4
and modify_field.p4 with some fixes.