DPMU: Add commands file support; P4C-HP4: fix match template bug
In the DPMU, the --commands option is now supported which allows
the client to specify a file containing multiple rules to be
translated and pushed to HP4.

In p4c-hp4, a bug was producing extra match templates, an artifact
of the way standard metadata field matching used to require
two tables.  This bug has been fixed.