p4c-hp4: fix ingress pipeline parsing (walk_ingress_pipeline)
As noted in previous commit, walk_ingress_pipeline was employing
an algorithm appropriate for traversing a tree, but it is not
safe to assume the source program's control flow is a tree.
It is a DAG, and this commit changes walk_ingress_pipeline to
do a topological sort on the DAG to determine stage assignments.

The result was tested vs. arp_proxy.p4 and looks good.