DPMU: Fix handling of match params; general: ID bug
DPMU was not handling hex-formatted match parameters in received
rules but now it is.

Proceeded with arp_proxy test and we got as far as t_mod_41 or
the first primitive in the 'arp_reply' action in arp_proxy.p4
before we got an unexpected TABLE_MISS.  Further diagnosis
required.

But we are really dang close to getting arp_proxy demo working.