Reorg, replace term DPMU w/ controller; start new implementation
New implementation will use what code we can from the old, but
there are some fundamental organization differences as a result
of design work done this week, especially with respect to
support for compositions (see hp4-compilers repo, design.tex
and compositions.md).

This commit includes partial solution for the "insert" operation
in the UDev class.  It also includes numerous other classes
required by the controller as identified in the design.