Rework parse optimizations, no change in performance
Thought avoiding resubmits would be key to significant
increase in throughput, but this commit casts doubt
on the idea that resubmits are the main contributor
to the throughput penalty.  The next obvious thing
is the number of actions that have dependencies
among their primitives, which would induce stalls
in the pipeline.