Fix next_table field width to handle MATCHLESS value (99)
Also, comments added to help identify conditionals, supporting
debugging via the event logger.

The next_table field was 4 bits, but init_program_state was
assigning a value of '99' corresponding to matchless table type.
When a modify_field tries to assign a value larger than the width
of the target, bmv2 silently fails.

Over in hp4-ctrl, t06 now works in its most essential test -
a chain is created with simple_router -> firewall, and the
router routes, and the firewall blocks/doesn't block correctly!