Confirm conditions for running tests / no unwanted output
Unwanted output noted in previous commit was result of forgetting
that the hp4controller modules had been installed to /usr/local/...
via setuptools (setup.py in hp4-ctrl), so the modules being
accessed were not the ones I was modifying in my home directory.

To more easily modify these files and run tests, instead of
using setuptools we modify PYTHONPATH in .profile in the home
directory.  We must also modify sudoers to preserve the
environment when running commands via sudo, because the
test requires the use of sudo, because of mininet.

We should add basic instructions to this effect.